feat: value validations

+++ fruits/database/index.js
@@ -16,6 +16,11 @@ app.get('/fruits', async (req, res) => {
 
 app.get('/fruits/:id', async (req, res) => {
     const { id } = req.params;
+
+    if(id < 0){
+        return  res.status(400).json({ error: 'Az ID nem lehet negatív szám !' });
+    }
+
     try {
         const [rows] = await fruit.query('SELECT * FROM fruits WHERE id = ?', [id]);
         res.json(rows);
@@ -26,6 +31,14 @@ app.get('/fruits/:id', async (req, res) => {
 
 app.post('/fruits', express.json(), async (req, res) => {
     const { name, color, price } = req.body;
+
+    if(price < 0){
+        return res.status(400).json({ error: 'Az ár nem lehet negatív szám !' });
+    }
+    if(name, color != String){
+        return res.status(400).json({ error: 'A név és a szín csak szöveg lehet !' });
+    }
+
     try {
         const [result] = await fruit.query('INSERT INTO fruits (name, color, price) VALUES (?, ?, ?)', [name, color, price]);
         res.status(201).json({ id: result.insertId, name, color, price });
@@ -37,6 +50,17 @@ app.post('/fruits', express.json(), async (req, res) => {
 app.put('/fruits/:id', express.json(), async (req, res) => {
     const { id } = req.params;
     const { name, color, price } = req.body;
+
+    if(id < 0){
+        return  res.status(400).json({ error: 'Az ID nem lehet negatív szám !' });
+    }
+    if(price < 0){
+        return res.status(400).json({ error: 'Az ár nem lehet negatív szám !' });
+    }
+    if(name, color != String){
+        return res.status(400).json({ error: 'A név és a szín csak szöveg lehet !' });
+    }
+
     try {
         const [result] = await fruit.query('UPDATE fruits SET name = ?, color = ?, price = ? WHERE id = ?', [name, color, price, id]);
         if (result.affectedRows === 0) {
@@ -51,6 +75,11 @@ app.put('/fruits/:id', express.json(), async (req, res) => {
 
 app.delete('/fruits/:id', async (req, res) => {
     const { id } = req.params;
+
+    if(id < 0){
+        return  res.status(400).json({ error: 'Az ID nem lehet negatív szám !' });
+    }
+
     try {
         const [result] = await fruit.query('DELETE FROM fruits WHERE id = ?', [id]);
         if (result.affectedRows === 0) {
